opensubtitles reactoring
