+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/subtitleproviders/opensubtitles/SearchSubtitles.java
@@ -1,16 +1,16 @@
-package org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2;
+package org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles;
 
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.exception.OpenSubtitlesException;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.AiTranslatedEnum;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.ForeignPartsOnlyEnum;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.HearingImpairedEnum;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.MachineTranslatedEnum;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.MoviehashMatchEnum;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.OrderDirectionEnum;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.ParamIntf;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.SearchSubtitlesEnum;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.TrustedSourcesEnum;
-import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param.TypeEnum;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.exception.OpenSubtitlesException;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.AiTranslatedEnum;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.ForeignPartsOnlyEnum;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.HearingImpairedEnum;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.MachineTranslatedEnum;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.MoviehashMatchEnum;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.OrderDirectionEnum;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.ParamIntf;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.SearchSubtitlesEnum;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.TrustedSourcesEnum;
+import org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param.TypeEnum;
 import org.lodder.subtools.sublibrary.Language;
 import org.lodder.subtools.sublibrary.Manager;
 import org.lodder.subtools.sublibrary.cache.CacheType;
