--- MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/subtitleproviders/opensubtitles/api/v2/param/ParamIntf.java
@@ -1,5 +1,0 @@
-package org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param;
-
-public interface ParamIntf {
-    String getValue();
-}
