+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/subtitleproviders/opensubtitles/param/MachineTranslatedEnum.java
@@ -1,4 +1,4 @@
-package org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param;
+package org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -6,10 +6,11 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-public enum OrderDirectionEnum implements ParamIntf {
-    ACSENDING("asc"), DESCENDING("desc");
+public enum MachineTranslatedEnum implements ParamIntf {
+    EXCLUDE("exclude"), INCLUDE("include");
 
     private final String value;
+
 
     @Override
     public String toString() {
