+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/subtitleproviders/opensubtitles/param/ParamIntf.java
@@ -1,0 +1,5 @@
+package org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param;
+
+public interface ParamIntf {
+    String getValue();
+}
