+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/subtitleproviders/opensubtitles/param/TypeEnum.java
@@ -1,4 +1,4 @@
-package org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.api.v2.param;
+package org.lodder.subtools.multisubdownloader.subtitleproviders.opensubtitles.param;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -6,11 +6,10 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-public enum MachineTranslatedEnum implements ParamIntf {
-    EXCLUDE("exclude"), INCLUDE("include");
+public enum TypeEnum implements ParamIntf {
+    MOVIE("movie"), EPISODE("episode"), ALL("all");
 
     private final String value;
-
 
     @Override
     public String toString() {
